% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mevolCVP.R
\name{mevol_CVP}
\alias{mevol_CVP}
\title{Do mevolCVP}
\usage{
mevol_CVP(
  mat = mat,
  group = group,
  lim = 30,
  nrep = 100,
  return.matrix = FALSE,
  minInd = NULL,
  print.legend = TRUE
)
}
\arguments{
\item{mat}{A PCA matrix (individuals in rows and variables in columns).}

\item{group}{the grouping factor (a vector or a factor)}

\item{lim}{maximum number of components to be included}

\item{nrep}{number of resampling.}

\item{return.matrix}{if =TRUE, the matrix used to draw the graphics is returned.}

\item{minInd}{the minimum number of specimens to be included in the analyses. By default the number of specimens in the smallest group.}

\item{print.legend}{\code{logical} whether to add a legend}
}
\value{
a graph and if \code{return.matrix} is \code{TRUE}, then a \code{list}
}
\description{
More about it
}
\examples{
# Some examples
mevol_CVP(pig$mat, pig$gp, nrep=2)

}
