% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{mevolCVP-package}
\alias{mevolCVP}
\alias{mevolCVP-package}
\title{mevolCVP: Bla blah}
\description{
More blah blah
}
\author{
\strong{Maintainer}: Allowen Evin \email{allowen.evin@umontpellier.fr}

Authors:
\itemize{
  \item Vincent Bonhomme \email{vincent.bonhomme@umontpellier.fr}
}

}
\keyword{internal}
