% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pig.R
\docType{data}
\name{pig}
\alias{pig}
\title{Data from JAS}
\format{A \code{list} containing \code{gp} (a factor) and \code{mat} (a \code{matrix} of dim 171x171).}
\source{
See citation("mevolCVP")
}
\description{
Data from JAS
}
