% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pldam.R
\name{pldam}
\alias{pldam}
\title{Predictive discriminant analyses based on balanced samples.}
\usage{
pldam(mat = mat, group = group, mattoid = mattoid, codetoid = codetoid,
  ncpts = 20, limproba = 0, limCVP = FALSE, nrep = 100,
  return.matrix = TRUE, minInd = NULL)
}
\arguments{
\item{mat}{matrix with the 'known' data (usually PC scores, but works also with univariate data)}

\item{group}{grouping factor of the 'known' data}

\item{mattoid}{matrix with the data to be predicted}

\item{codetoid}{label of the specimens to predict}

\item{ncpts}{number of PCs to retain}

\item{limproba}{a limit for the posterior probability to be concidered}

\item{limCVP}{a limit for the cross validation percent of the discriminate analysis to be concidered}

\item{nrep}{number of replicates}

\item{return.matrix}{\code{logical} whether to return full results}

\item{minInd}{the number of specimens to include for the balanced design if smaller than the smallest of the "known" groups}
}
\value{
a \code{list}
}
\description{
More about it
}
\examples{
# Some examples
# Grab some data from pig
mat <- pig$mat
gp <- pig$gp
# For the sake of reproducibility
set.seed(123)
train_ids <- sample(nrow(mat), 150, replace=FALSE)

pldam(# train data
      mat=mat[train_ids, ], group=gp[train_ids],
      # test data
      mattoid=mat[-train_ids, ], codetoid=paste0("id_", 1:nrow(mat[-train_ids, ])),
      nrep=2)
}
